+++ src/controllers/hellomeg/index.test.ts
@@ -1,14 +1,14 @@
 import { describe, expect, it, vi, beforeEach } from "vitest";
 
-import type { Env } from "../../util/env";
-import { Rarity } from "../../service/gacha";
-import { helloruriController } from "./index";
-import * as gachaService from "../../service/gacha";
-import { HELLORURI_MESSAGE_NORMAL, HELLORURI_MESSAGE_UR } from "./constants";
+import type { Env } from "../../utils/env";
+import { Rarity } from "../../services/gacha";
+import { hellomegController } from "./index";
+import * as gachaService from "../../services/gacha";
+import { HELLOMEG_MESSAGE_NORMAL, HELLOMEG_MESSAGE_UR } from "./constants";
 
-vi.mock("../../service/gacha", async () => {
-  const actual = await vi.importActual<typeof import("../../service/gacha")>(
-    "../../service/gacha",
+vi.mock("../../services/gacha", async () => {
+  const actual = await vi.importActual<typeof import("../../services/gacha")>(
+    "../../services/gacha",
   );
   return {
     ...actual,
@@ -22,8 +22,8 @@ function createEnv(overrides?: Partial<Env>): Env {
     ASSETS: { fetch: vi.fn() },
     ASSETS_BASE_URL: "https://cdn.example.com",
     DISCORD_PUBLIC_KEY: "pub",
-    HELLORURI_UR_PROBABILITY: "0.11",
-    HELLORURI_SR_PROBABILITY: "0.22",
+    HELLOMEG_UR_PROBABILITY: "0.11",
+    HELLOMEG_SR_PROBABILITY: "0.22",
     ...overrides,
   };
 }
@@ -33,17 +33,17 @@ beforeEach(() => {
   vi.spyOn(console, "error").mockImplementation(() => {});
 });
 
-describe("helloruriController", () => {
-  it("returns JSON with UR message", async () => {
+describe("hellomegController", () => {
+  it("returns JSON with UR large message", async () => {
     vi.mocked(gachaService.drawRarity).mockReturnValue(Rarity.UR);
 
     const fetchSpy = vi.fn();
     vi.stubGlobal("fetch", fetchSpy);
 
-    const response = await helloruriController(createEnv());
+    const response = await hellomegController(createEnv());
     const payload = await response.json();
 
-    expect(payload.data.content).toBe(HELLORURI_MESSAGE_UR);
+    expect(payload.data.content).toBe(HELLOMEG_MESSAGE_UR);
     expect(response.headers.get("content-type")).toContain("application/json");
     expect(gachaService.drawRarity).toHaveBeenCalledWith(0.11, 0.22);
     expect(gachaService.buildSrResult).not.toHaveBeenCalled();
@@ -54,17 +54,17 @@ describe("helloruriController", () => {
     vi.mocked(gachaService.drawRarity).mockReturnValue(Rarity.SR);
     vi.mocked(gachaService.buildSrResult).mockResolvedValue(null);
 
-    const response = await helloruriController(createEnv());
+    const response = await hellomegController(createEnv());
     const payload = await response.json();
 
-    expect(payload.data.content).toBe(HELLORURI_MESSAGE_NORMAL);
+    expect(payload.data.content).toBe(HELLOMEG_MESSAGE_NORMAL);
     expect(response.headers.get("content-type")).toContain("application/json");
   });
 
   it("returns multipart response when SR image fetch succeeds", async () => {
     vi.mocked(gachaService.drawRarity).mockReturnValue(Rarity.SR);
     vi.mocked(gachaService.buildSrResult).mockResolvedValue({
-      filepath: "public/helloruri/sr/sample.png",
+      filepath: "public/hellomegbot/sr/sample.png",
       twitterId: "artist_id",
     });
 
@@ -75,12 +75,12 @@ describe("helloruriController", () => {
     );
     vi.stubGlobal("fetch", fetchSpy);
 
-    const response = await helloruriController(createEnv());
+    const response = await hellomegController(createEnv());
     const bodyText = new TextDecoder().decode(await response.arrayBuffer());
 
     expect(response.headers.get("content-type")).toContain("multipart/form-data");
     expect(fetchSpy).toHaveBeenCalledWith(
-      "https://cdn.example.com/public/helloruri/sr/sample.png",
+      "https://cdn.example.com/public/hellomegbot/sr/sample.png",
     );
     expect(bodyText).toContain('name="payload_json"');
     expect(bodyText).toContain("イラスト：[@artist_id](<https://twitter.com/artist_id>)");
@@ -90,13 +90,13 @@ describe("helloruriController", () => {
   it("falls back to JSON when SR image fetch fails", async () => {
     vi.mocked(gachaService.drawRarity).mockReturnValue(Rarity.SR);
     vi.mocked(gachaService.buildSrResult).mockResolvedValue({
-      filepath: "public/helloruri/sr/sample.png",
+      filepath: "public/hellomegbot/sr/sample.png",
       twitterId: "artist_id",
     });
 
     vi.stubGlobal("fetch", vi.fn().mockRejectedValue(new Error("boom")));
 
-    const response = await helloruriController(createEnv());
+    const response = await hellomegController(createEnv());
     const payload = await response.json();
 
     expect(payload.data.content).toBe(
